I did a lot of css tuning to get it to look like the business doc example. I also rearranged my file structure, just moved my modules out of their sub-folders and into the main widget folder.
