+++ client/src/components/related_products/Modal.jsx
@@ -5,8 +5,8 @@ export default function Modal({ open, onClose, children }) {
   if (!open) return null;
   return (
     <>
-      <div className="overlay" />
-      <div className="modal">
+      <div className="overlay" role="presentation" onClick={onClose} />
+      <div className="modal1">
         <div className="modal-title">
           Comparison
         </div>
