+++ client/src/components/related_products/ProductCard.css
@@ -1,22 +1,21 @@
 .cardContainer {
   background-color: rgb(232, 232, 232);
-  height: 100%;
+  height: 99%;
   min-width: 220px;
   max-width: 220px;
   border: solid;
   border-width: 2px;
-
   border-color: rgb(45, 45, 45);
   cursor: pointer;
   z-index: 0;
   position: relative;
+  box-shadow: rgba(0, 0, 0, 0.3) 0px 4px 8px 0px;
 }
 
 .imgDiv{
   height: 72%;
   width: 100%;
   margin: auto;
-
   margin-bottom: 5px;
   background-color: rgb(114, 114, 114);
   position: relative;
@@ -34,15 +33,19 @@
 
 .compareButton {
   position: absolute;
-  top: 3px;
-  right: 3px;
+  top: 10px;
+  right: 10px;
   z-index: 5;
   border: none;
-  border-radius: 15px;;
+  border-radius: 20px;;
   background: rgba(0,0,0,.6);
   font-size: 20px;
   color: gold;
   cursor: pointer;
+  padding-bottom: 3px;
+  width: 33px;
+  height: 33px;
+  box-shadow: rgba(0, 0, 0, 0.3) 0px 4px 8px 0px;
 }
 
 .category {
