+++ client/src/components/related_products/Related.css
@@ -1,25 +1,24 @@
-body {
-
-}
-
 .widgetContainer {
-  width: 880px;
+  width: 900px;
   display: flex;
   flex-direction: column;
   align-items: center;
-  background-color: rgb(215, 215, 215);
-  height: 800px;
+  height: 824px;
 }
 
 .closeModalBtn {
   background-color: rgba(0, 0, 0, .4);
   border: none;
-  border-radius: 5px;
+  border-radius: 50px;
   width: 30%;
-  margin: 10px;
+  margin-top: 25px;
+  height: 40px;
+  font-size: 18px;
+  background-color: rgb(220, 220, 220);
+  font-weight: 400;
 }
 
-.modal {
+.modal1 {
   position: fixed;
   top: 50%;
   left: 50%;
@@ -27,6 +26,11 @@ body {
   padding: 40px;
   border-radius: 10px;
   height: fit-content;
+  z-index: 10;
+  background-color: rgb(255, 255, 255);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
 
 .overlay {
@@ -41,9 +45,14 @@ body {
 
 .modal-title {
   color: black;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
+  width: fit-content;
+  background-color: rgb(220, 220, 220);
+  padding-top: 15px;
+  padding-bottom: 15px;
+  padding-left: 40px;
+  padding-right: 40px;
+  border-radius: 5px;
+  font-weight: 700;
 }
 
 .modal-content {
@@ -58,33 +67,28 @@ table {
   border: 1px solid black;
   table-layout: fixed;
   margin-top: 20px;
-}
-
-table td {
-  border: 1px solid black;
-  text-align: center;
+  border-collapse: collapse;
 }
 
 tbody {
   vertical-align: baseline;
   align-items: center;
   width: 100%;
-
 }
 
 tr {
   width: 100%;
-
+  height: 40px;
+  text-align: center;
 }
 
-th.currentP {
-
+th {
+  vertical-align: middle;
+  border: 1px solid black;
 }
 
-th.feature {
-
-}
-
-th.compareP {
-
+td {
+  border: 1px solid black;
+  text-align: center;
+  vertical-align: middle;
 }
