+++ client/src/components/related_products/__test__/Carousel.spec.js
@@ -5,8 +5,8 @@ import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import { act } from 'react-test-renderer';
 import '@testing-library/jest-dom';
-import ProductCard from '../ProductCard.jsx';
-import { GlobalContext } from '../../../GlobalContext.jsx';
+import Carousel from '../Carousel.jsx';
+import { GlobalContext } from '../../GlobalContext.jsx';
 
 describe('Carousel component', () => {
   describe('Related Products Carousel', () => {
@@ -16,7 +16,7 @@ describe('Carousel component', () => {
           {
             name: 'Blank',
             slogan: 'No product yet',
-            category: 'outfitBlank',
+            category: 'outfit placeholder',
             default_price: '$$',
           },
           {
@@ -41,10 +41,70 @@ describe('Carousel component', () => {
           },
         ],
 
+        40345: [
+          {
+            name: 'Sunglasses',
+            slogan: 'Protect your eyes',
+            category: 'Accessories',
+            default_price: '999.00',
+          },
+          {
+            results: [
+              {
+                'default?': true,
+                photos: [
+                  {
+                    thumbnail_url: null,
+                  },
+                ],
+              },
+            ],
+          },
+          {
+            ratings: {
+              1: '0',
+              2: '0',
+              3: '0',
+              4: '1',
+              5: '0',
+            },
+          },
+        ],
+
+        40346: [
+          {
+            name: 'Joggers',
+            slogan: 'Protect your legs',
+            category: 'Pants',
+            default_price: '998.00',
+          },
+          {
+            results: [
+              {
+                'default?': true,
+                photos: [
+                  {
+                    thumbnail_url: 'https://images.unsplash.com/photo-1552902865-b72c031ac5ea?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=300&q=80',
+                  },
+                ],
+              },
+            ],
+          },
+          {
+            ratings: {
+              1: '0',
+              2: '0',
+              3: '1',
+              4: '0',
+              5: '0',
+            },
+          },
+        ],
+
         40350: [
           {
             name: 'Blues Shoes',
-            slogan: 'Protect your feet and this slogan is super long to make sure it gets cut down by my expanded title function',
+            slogan: 'Protect your feet',
             category: 'Shoes',
             default_price: '997.00',
           },
@@ -133,21 +193,20 @@ describe('Carousel component', () => {
       };
       const mockProductID = 40344;
       const yes = true;
-      const relatedID = 40351;
-      let burn = 0;
-      const mockToggle = () => {
-        burn = 1;
-        console.log('triggered');
-      };
+      const related = [40344, 40345, 40346, 40351, 40350];
+      const burn = 0;
+      const outfitList = [];
+      const mockSetID = () => { console.log('id change'); };
       await act(() => {
         render(
-          <GlobalContext.Provider value={{ productID: mockProductID }}>
-            <ProductCard
+          <GlobalContext.Provider value={{ productID: mockProductID, setProductID: mockSetID }}>
+            <Carousel
               rpMode={yes}
               dataStore={dataStore}
-              relatedID={relatedID}
+              related={related}
               burn={burn}
-              triggerFunction={mockToggle}
+              outfitList={outfitList}
+              outfitToggle={() => {}}
             />
           </GlobalContext.Provider>,
         );
@@ -158,58 +217,49 @@ describe('Carousel component', () => {
       expect(screen.getByText('Kicks')).toBeVisible();
     });
 
-    it('should not render wrong product category', () => {
-      expect(screen.queryByText('Jackets')).toBe(null);
-    });
-
     it('should render product name', () => {
-      expect(screen.getByText(/Yeasy 350/i)).toBeVisible();
-    });
-
-    it('should render expanded product name', () => {
-      expect(screen.getByText(/Yeasy 350 - Protect your feets/)).toBeVisible();
+      expect(screen.getByText(/Joggers/i)).toBeVisible();
     });
 
     it('should render product price', () => {
-      expect(screen.getByText(/996.00/i)).toBeVisible();
+      expect(screen.getByText(/999.00/i)).toBeVisible();
     });
 
     it('should render product slogan', () => {
-      expect(screen.getByText(/Protect your feets/i)).toBeVisible();
+      expect(screen.getByText(/Protect your legs/i)).toBeVisible();
     });
 
-    it('should render stars', () => {
-      expect(screen.getByText('★★★★★')).toBeVisible();
+    it('should render nav buttons if product list > 4', () => {
+      expect(screen.getByText(/>/i)).toBeVisible();
     });
 
-    it('should trigger changeProduct function when card is clicked', async () => {
+    it('should call slide function when nav button is clicked', async () => {
       const logSpy = jest.spyOn(console, 'log');
 
-      const productCard = screen.getByText(/Kicks/i);
-      const productCardBtn = productCard.closest('.cardContainer');
-      await fireEvent.click(productCardBtn);
+      const navBtn = screen.getByText(/>/i);
+      await fireEvent.click(navBtn);
 
-      expect(logSpy).toHaveBeenCalledWith('triggered');
+      expect(logSpy).toHaveBeenCalledWith('RP slide');
     });
 
-    it('should trigger compare buttons console.log when clicked', async () => {
+    it('should call changeProduct function when card is clicked', async () => {
       const logSpy = jest.spyOn(console, 'log');
 
-      const compareBtn = screen.getByText('★');
-      await fireEvent.click(compareBtn);
+      const productCard = screen.getByText('Kicks');
+      await fireEvent.click(productCard);
 
-      expect(logSpy).toHaveBeenCalledWith('compare button pressed');
+      expect(logSpy).toHaveBeenCalledWith('id change');
     });
   });
 
-  describe('Related Products Carousel', () => {
+  describe('Outfit Carousel', () => {
     beforeEach(async () => {
       const dataStore = {
         10001: [
           {
             name: 'Blank',
             slogan: 'No product yet',
-            category: 'outfitBlank',
+            category: 'outfit placeholder',
             default_price: '$$',
           },
           {
@@ -234,17 +284,77 @@ describe('Carousel component', () => {
           },
         ],
 
+        40345: [
+          {
+            name: 'Sunglasses',
+            slogan: 'Protect your eyes',
+            category: 'Accessories',
+            default_price: '999.00',
+          },
+          {
+            results: [
+              {
+                'default?': true,
+                photos: [
+                  {
+                    thumbnail_url: null,
+                  },
+                ],
+              },
+            ],
+          },
+          {
+            ratings: {
+              1: '0',
+              2: '0',
+              3: '0',
+              4: '1',
+              5: '0',
+            },
+          },
+        ],
+
+        40346: [
+          {
+            name: 'Joggers',
+            slogan: 'Protect your legs',
+            category: 'Pants',
+            default_price: '998.00',
+          },
+          {
+            results: [
+              {
+                'default?': true,
+                photos: [
+                  {
+                    thumbnail_url: 'https://images.unsplash.com/photo-1552902865-b72c031ac5ea?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=300&q=80',
+                  },
+                ],
+              },
+            ],
+          },
+          {
+            ratings: {
+              1: '0',
+              2: '0',
+              3: '1',
+              4: '0',
+              5: '0',
+            },
+          },
+        ],
+
         40350: [
           {
             name: 'Blues Shoes',
-            slogan: 'Protect your feet and this slogan is super long to make sure it gets cut down by my expanded title function',
+            slogan: 'Protect your feet',
             category: 'Shoes',
             default_price: '997.00',
           },
           {
             results: [
               {
-                'default?': false,
+                'default?': true,
                 photos: [
                   {
                     thumbnail_url: 'https://images.unsplash.com/photo-1561861422-a549073e547a?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=300&q=80',
@@ -325,22 +435,19 @@ describe('Carousel component', () => {
         ],
       };
       const mockProductID = 40344;
-      const yes = true;
-      const relatedID = 40350;
-      let burn = 0;
-      const mockToggle = () => {
-        burn = 1;
-        console.log('triggered');
-      };
+      const related = [40344, 40345, 40346, 40351, 40350];
+      const burn = 0;
+      const outfitList = [10001, 10001, 10001];
       await act(() => {
         render(
           <GlobalContext.Provider value={{ productID: mockProductID }}>
-            <ProductCard
-              rpMode={yes}
+            <Carousel
+              rpMode={false}
               dataStore={dataStore}
-              relatedID={relatedID}
+              related={related}
               burn={burn}
-              triggerFunction={mockToggle}
+              outfitList={outfitList}
+              outfitToggle={() => {}}
             />
           </GlobalContext.Provider>,
         );
@@ -348,50 +455,27 @@ describe('Carousel component', () => {
     });
 
     it('should render product category', () => {
-      expect(screen.getByText('Shoes')).toBeVisible();
-    });
-
-    it('should not render wrong product category', () => {
-      expect(screen.queryByText('Jackets')).toBe(null);
+      expect(screen.getByText('Jackets')).toBeVisible();
     });
 
     it('should render product name', () => {
-      expect(screen.getByText(/Blues Shoes/i)).toBeVisible();
+      expect(screen.getByText(/Camo Onsie/i)).toBeVisible();
     });
 
-    it('should render expanded product name', () => {
-      expect(screen.getByText(/Blues Shoes - Protect your feet and this slog.../)).toBeVisible();
+    it('should render product name', () => {
+      expect(screen.getAllByText(/Blank/i)).toHaveLength(3);
     });
 
     it('should render product price', () => {
-      expect(screen.getByText(/997.00/i)).toBeVisible();
+      expect(screen.getByText(/995.00/i)).toBeVisible();
     });
 
     it('should render product slogan', () => {
-      expect(screen.getByText(/Protect your feet and/i)).toBeVisible();
+      expect(screen.getByText(/Protect your torso/i)).toBeVisible();
     });
 
-    it('should render stars', () => {
-      expect(screen.getByText('★★★★★')).toBeVisible();
-    });
-
-    it('should trigger changeProduct function when card is clicked', async () => {
-      const logSpy = jest.spyOn(console, 'log');
-
-      const productCard = screen.getByText(/Blues Shoes/i);
-      const productCardBtn = productCard.closest('.cardContainer');
-      await fireEvent.click(productCardBtn);
-
-      expect(logSpy).toHaveBeenCalledWith('triggered');
-    });
-
-    it('should trigger compare buttons console.log when clicked', async () => {
-      const logSpy = jest.spyOn(console, 'log');
-
-      const compareBtn = screen.getByText('★');
-      await fireEvent.click(compareBtn);
-
-      expect(logSpy).toHaveBeenCalledWith('compare button pressed');
+    it('should render blank outfit cards', () => {
+      expect(screen.getAllByText('$$$')).toHaveLength(3);
     });
   });
 });
