--- client/src/components/related_products/row2/OutfitCard.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function OutfitCard() {
-  return (
-    <div>
-      outfit card
-    </div>
-  );
-}
-
-export default OutfitCard;
