--- client/src/components/related_products/row2/RowTwo.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function RowTwo() {
-  return (
-    <div>
-      row two
-    </div>
-  );
-}
-
-export default RowTwo;
